refactored naming, formatting fixes, published CoralInAndOut state
