+++ src/main/java/org/team5924/frc2025/BuildConstants.java
@@ -21,12 +21,12 @@ public final class BuildConstants {
   public static final String MAVEN_GROUP = "";
   public static final String MAVEN_NAME = "GoldenGateRobotics2025";
   public static final String VERSION = "unspecified";
-  public static final int GIT_REVISION = 22;
-  public static final String GIT_SHA = "dbf711f330ae5fe4f7d0a81295f05f0fb13ecf5f";
-  public static final String GIT_DATE = "2025-01-25 18:49:52 EST";
+  public static final int GIT_REVISION = 23;
+  public static final String GIT_SHA = "df9669a054ae723fd0bc896266c54b9f8c7f7868";
+  public static final String GIT_DATE = "2025-01-25 19:31:12 EST";
   public static final String GIT_BRANCH = "3-coral-in-and-out";
-  public static final String BUILD_DATE = "2025-01-25 19:29:22 EST";
-  public static final long BUILD_UNIX_TIME = 1737851362718L;
+  public static final String BUILD_DATE = "2025-01-25 19:39:46 EST";
+  public static final long BUILD_UNIX_TIME = 1737851986231L;
   public static final int DIRTY = 1;
 
   private BuildConstants() {}
