+++ src/main/java/org/team5924/frc2025/subsystems/rollers/CoralInAndOut/CoralInAndOutIOKrakenFOC.java
@@ -14,7 +14,7 @@
  * If you did not, see <https://www.gnu.org/licenses>.
  */
 
-package org.team5924.frc2025.subsystems.rollers.intake;
+package org.team5924.frc2025.subsystems.rollers.CoralInAndOut;
 
 import au.grapplerobotics.LaserCan;
 import org.team5924.frc2025.Constants;
